[SOT] rewrite eval frame in C-lang (#57653)
