+++ test/dygraph_to_static/test_pylayer.py
@@ -14,13 +14,19 @@
 
 """Tests for PyLayer of Dynamic-to-Static.
 Only test simple cases here."""
+import sys
+from pathlib import Path
+
+sys.path.append(
+    str(Path(__file__).absolute().parent.parent.joinpath("legacy_test"))
+)
 
 import os
 import tempfile
 import unittest
 
 import numpy as np
-from legacy_test.test_jit_save_load import train
+from test_jit_save_load import train
 
 import paddle
 from paddle.autograd.py_layer import PyLayer
